+++ music/src/Service/Uploader/AvatarUploader.php
@@ -36,7 +36,9 @@ class AvatarUploader
         $path = 'avatar/' . $userId;
         $name = Uuid::uuid4()->toString() . '.' . $file->getClientOriginalExtension();
 
-        $this->storage->deleteDir($path);
+        if (file_exists($path)) {
+            $this->storage->deleteDir($path);
+        }
         $this->storage->createDir($path);
         $stream = fopen($file->getRealPath(), 'rb+');
         $this->storage->writeStream($path . '/' . $name, $stream);
